requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework>=3.0
python-dotenv==0.19.2
flake8==4.0.1
django-filter==21.1
